pyproject.toml
[tool.poetry]
name = "rest api contacts"
version = "0.1.0"
description = ""
authors = ["Yurii-Kovalenko <yuriy.kovalenko.in@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
redis = "^5.0.3"
fastapi = "^0.110.1"
fastapi-limiter = "^0.1.6"
uvicorn = "^0.29.0"
sphinx = "^7.2.6"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.29"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
cloudinary = "^1.39.1"
fastapi-mail = "^1.4.1"
psycopg2 = "^2.9.9"
python-multipart = "^0.0.9"
pytest = "^8.1.1"
httpx = "^0.27.0"
alembic = "^1.13.1"


[tool.poetry.group.dev.dependencies]
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = ["."]
